Finalização das migrations iníciais

+++ api/src/main/resources/db/migration/V1__createTables.sql
@@ -1,32 +1,32 @@
 CREATE TABLE "Priorities" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "priority" VARCHAR(50) UNIQUE NOT NULL
 );
 
 CREATE TABLE "Status" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "status" VARCHAR(50) UNIQUE NOT NULL
 );
 
 CREATE TABLE "Classifications" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "classification" VARCHAR(50) UNIQUE NOT NULL
 );
 
 CREATE TABLE "Clients" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "client" VARCHAR(50) UNIQUE NOT NULL
 );
 
 CREATE TABLE "Projects" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "client_id" BIGINT,
   "team_id" BIGINT,
   "project" VARCHAR(50) UNIQUE NOT NULL
 );
 
 CREATE TABLE "Users" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "name" VARCHAR(50) NOT NULL,
   "last_name" VARCHAR(50),
   "email" VARCHAR(100) UNIQUE NOT NULL,
@@ -36,19 +36,19 @@ CREATE TABLE "Users" (
 );
 
 CREATE TABLE "Teams" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "name" VARCHAR(50) UNIQUE NOT NULL,
   "leader_id" BIGINT NOT NULL
 );
 
 CREATE TABLE "TeamPeople" (
-  "team_id" BIGINT,
-  "user_id" BIGINT,
+  "team_id" INTEGER,
+  "user_id" INTEGER,
   PRIMARY KEY ("team_id", "user_id")
 );
 
 CREATE TABLE "Tasks" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "status_id" BIGINT NOT NULL,
   "classification_id" BIGINT NOT NULL,
   "priority_id" BIGINT NOT NULL,
@@ -62,7 +62,7 @@ CREATE TABLE "Tasks" (
 );
 
 CREATE TABLE "Backlogs" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "note" VARCHAR,
   "created_at" TIMESTAMP,
   "task_id" BIGINT,
@@ -70,14 +70,14 @@ CREATE TABLE "Backlogs" (
 );
 
 CREATE TABLE "BacklogNotifications" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "notified" BOOLEAN,
   "backlog_id" BIGINT,
   "notified_user" BIGINT
 );
 
 CREATE TABLE "DetailedHours" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "description" VARCHAR(100),
   "time" TIME,
   "created_at" TIMESTAMP,
@@ -86,7 +86,7 @@ CREATE TABLE "DetailedHours" (
 );
 
 CREATE TABLE "Attachment" (
-  "id" BIGINT PRIMARY KEY,
+  "id" SERIAL PRIMARY KEY,
   "task_id" BIGINT,
   "name" VARCHAR(100),
   "path" VARCHAR(255),
